Restructure project
- use src layout
- split bcferries package into multiple modules
