+++ .gitignore
@@ -4,7 +4,7 @@ __pycache__/
 [Bb]uild/
 dist/
 *env*
-cache*
+cache*/
 nicepage/
 raw*/
 scripts/
