--- bcferries/__init__.py
@@ -1,15 +1,0 @@
-from .bcferries import (
-    find_routes,
-    load_data,
-    locations,
-    make_route_plans,
-    schedule_cache,
-)
-
-__all__ = [
-    "find_routes",
-    "load_data",
-    "locations",
-    "make_route_plans",
-    "schedule_cache",
-]
