+++ src/server.py
@@ -1,35 +1,37 @@
 from __future__ import annotations
 
 import asyncio
+from contextlib import asynccontextmanager
 from pathlib import Path
+from typing import TYPE_CHECKING
 
-from bcferries import (
-    find_routes,
-    load_data,
-    locations,
-    make_route_plans,
-    schedule_cache,
-)
-from bcferries.classes import (
-    FerrySchedule,
-    RoutePlan,
-    RoutePlansOptions,
-    ScheduleOptions,
-)
 from fastapi import FastAPI, HTTPException, Request, status
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 
-app = FastAPI()
-app.mount("/static", StaticFiles(directory="static"), name="static")
-templates = Jinja2Templates(directory="templates")
+from .data import load_data, locations
+from .route import RoutePlan, find_routes, make_route_plans
+from .schedule import FerrySchedule, ScheduleCache
+
+if TYPE_CHECKING:
+    from collections.abc import AsyncGenerator
+
+    from .options import RoutePlansOptions, ScheduleOptions
 
 
-@app.on_event("startup")
-async def startup() -> None:
+@asynccontextmanager
+async def lifespan(_: FastAPI) -> AsyncGenerator[None, None]:
     load_data(Path("data/data.json"))
     schedule_cache.start_refresh_thread()
+    yield
+
+
+app = FastAPI(lifespan=lifespan)
+ROOT_DIR = Path(__file__).parent
+app.mount("/static", StaticFiles(directory=ROOT_DIR / "static"), name="static")
+templates = Jinja2Templates(directory=ROOT_DIR / "templates")
+schedule_cache = ScheduleCache()
 
 
 @app.get("/", response_class=HTMLResponse)
@@ -60,7 +62,7 @@ async def api_schedule(options: ScheduleOptions) -> FerrySchedule | None:
 @app.post("/api/routeplans", response_model=list[RoutePlan])
 async def api_routeplans(options: RoutePlansOptions) -> list[RoutePlan]:
     routes = find_routes(options.origin, options.destination)
-    route_plans = make_route_plans(routes, options)
+    route_plans = make_route_plans(routes, options, schedule_cache.get)
     route_plans.sort(key=lambda plan: plan.duration)
     return route_plans
