+++ src/static/index.js
