+++ src/static/notifications.js
